bugfix(etl.transform): per-symbol override of time_shift_ms not working
BRENT.CMD-USD needs custom time_shift_ms in order to be fully correct.
Custom symbol override was not working because of a bug in transform.py.

I shifted BRENT.CMD-USD by 3 hours and now the 1h, 4h and 1d look exact.

Configuring this pipeline for symbols is a bit (too) tricky. Will think
of a way to make it easier. Important bugfix.