feature(config): added includes support
YAML configuration can now be parted through inclusion patterns

Example:

resample:
  round_decimals: 8                   # Number of decimals to round OHLCV to
  batch_size: 250000                  # Maximum number of lines to read per batch
  paths:
    data: data/resample               # Output directory for resampled timeframes
  timeframes:
    includes:
    - config/resample/timeframes.default.yaml
  # Support per symbol overrides
  symbols:
    includes:
    - config/resample/timeframes.bonds.yaml
    - config/resample/timeframes.commodities.yaml
    - config/resample/indices/*.yaml

How it works? It first loads the YAML and then recursively traverses it to find "includes"-keys.
It resolves each item of the list to its corresponding filenames and loads them and adds it to a dict
structure. Next, it export the resulting dict to a YAML.

The original YAML config loader is then used to load the YAML. This way we only needed to make
as little modifications as possible. Direct specification (the old way) stays compatible.