feature(resample): fix for SGD and more!
This is a major overhaul, also on the config part. We now support
per attribute overrides for Timeframes. So we dont need to specify attributes
that don't change with respect to the defaults anymore.

This makes config files a lot cleaner.

Also introduced a first version of the merge support. This fixes SGD.

It's still a bit ugly and hacky, but works